Create README.md
# Final-Year-Project
Final Year Project 
Project name : Triends
Project Description 
Triends is a travel web application for users who travel on their own or want to meet new people and make new friends on their journey.   
New people will be found by users’ preference setting and their profile which is their own setting for other users.
Not only you find the people but also the app has a functionality to send a message to friends whom you would like to spend time with for a specific event. 
One of the functions that the app has is displaying local events around the users in 5 km that users can enjoy in wherever they are.